Remove square brackets
Square brackets in model/reactions/species identifiers are not SBML-complient and do therefore not work with
cobrapy. Therefore, the following curations were necessary:
 - Changed ID metabolite endings from e.g. [e] to _e or [car] to _car
 - Removed all other square brackets from IDs
After doing this it turned out that the metabolite
Dodecanoyl_acyl_carrier_protein_c was duplicated (one with square
brackets). Removed the duplicated metabolite.

Additional: Gave name to compartments car and th, in addition to model
description